+++ src/fuzzy/sets/__init__.py
@@ -1,3 +1,19 @@
+"""
+Fuzzy sets module.
+"""
+
 from .abstract import FuzzySet
-from .impl import Triangular, LogGaussian, Gaussian, Lorentzian, LogisticCurve
 from .group import FuzzySetGroup
+from .membership import Membership
+from .impl import Triangular, LogGaussian, Gaussian, Lorentzian, LogisticCurve
+
+__all__ = [
+    "FuzzySet",
+    "FuzzySetGroup",
+    "Triangular",
+    "LogGaussian",
+    "Gaussian",
+    "Lorentzian",
+    "LogisticCurve",
+    "Membership",
+]
