--- src/fuzzy/sets/utils.py
@@ -1,28 +1,0 @@
-"""
-Utility functions, such as for getting all the subclasses of a given class.
-"""
-
-import inspect
-from typing import Dict, Any
-
-
-def get_object_attributes(obj_instance) -> Dict[str, Any]:
-    """
-    Get the attributes of an object instance.
-    """
-    # get the attributes that are local to the class, but may be inherited from the super class
-    local_attributes = inspect.getmembers(
-        obj_instance,
-        lambda attr: not (inspect.ismethod(attr)) and not (inspect.isfunction(attr)),
-    )
-    # get the attributes that are inherited from (or found within) the super class
-    super_attributes = inspect.getmembers(
-        obj_instance.__class__.__bases__[0],
-        lambda attr: not (inspect.ismethod(attr)) and not (inspect.isfunction(attr)),
-    )
-    # get the attributes that are local to the class, but not inherited from the super class
-    return {
-        attr: value
-        for attr, value in local_attributes
-        if (attr, value) not in super_attributes and not attr.startswith("_")
-    }
