+++ src/fuzzy/utils/functions.py
@@ -2,8 +2,9 @@
 Utility functions for fuzzy-theory.
 """
 
+import inspect
 from pathlib import Path
-from typing import Set, Any
+from typing import Set, Any, Dict
 
 
 def check_path_to_save_torch_module(path: Path) -> None:
@@ -39,3 +40,25 @@ def all_subclasses(cls) -> Set[Any]:
         A set of all subclasses of the given class.
     """
     return {cls}.union(s for c in cls.__subclasses__() for s in all_subclasses(c))
+
+
+def get_object_attributes(obj_instance) -> Dict[str, Any]:
+    """
+    Get the attributes of an object instance.
+    """
+    # get the attributes that are local to the class, but may be inherited from the super class
+    local_attributes = inspect.getmembers(
+        obj_instance,
+        lambda attr: not (inspect.ismethod(attr)) and not (inspect.isfunction(attr)),
+    )
+    # get the attributes that are inherited from (or found within) the super class
+    super_attributes = inspect.getmembers(
+        obj_instance.__class__.__bases__[0],
+        lambda attr: not (inspect.ismethod(attr)) and not (inspect.isfunction(attr)),
+    )
+    # get the attributes that are local to the class, but not inherited from the super class
+    return {
+        attr: value
+        for attr, value in local_attributes
+        if (attr, value) not in super_attributes and not attr.startswith("_")
+    }
